requirements.txt
SQLAlchemy==1.3.19
discord.py==1.7.2
discord==1.0.1
python-dotenv==0.17.1
psycopg2-binary==2.8.6
